fix: :recycle: incorporated review comments
